+++ src/Schedulers.jl
@@ -27,8 +27,8 @@ Configuration parameters for `Scheduler`.
 - `lfreq`: frequency of writing iteration logs (0: none, >0: number of iterations, 
     -1: iteration 1,2,5,10,20,...)
 - `tciter`: maximum number of iterations without improvement (<0: turned off)
-- `ttime`: time limit [s] (<0: turned off)
-- `tctime`: maximum time [s] without improvement (<0: turned off)
+- `ttime`: time limit in seconds (<0: turned off)
+- `tctime`: maximum time in seconds without improvement (<0: turned off)
 - `tobj`: objective value at which should be terminated when reached (<0: turned off)
 - `checkit`: call `check` for each solution after each method application
 - `log`: if true write all log information, else none